Fixed some memory issues by adding a c++ style allocator that calls constructors

+++ src/Platform/platform.cpp
@@ -198,7 +198,7 @@ namespace BlitzenPlatform
 
         void PlatformSleep(uint64_t ms)
         {
-            Sleep(ms);
+            Sleep(static_cast<DWORD>(ms));
         }
 
         void CreateVulkanSurface(VkInstance& instance, VkSurfaceKHR& surface, VkAllocationCallbacks* pAllocator)
